Ensures resources are cleaned up from disk (#28)
* Ensures resources are cleaned up from disk

* Ensures file exists before deleting

* Ensures file exists before attempting to delete it

+++ Packages/VirtualMachine/Sources/VirtualMachineResourcesServiceEditor/VirtualMachineResourcesServiceEditor.swift
@@ -26,6 +26,8 @@ public struct VirtualMachineResourcesServiceEditor: VirtualMachineResourcesServi
     }
 
     public func removeResources() throws {
-        try fileSystem.removeItem(at: directoryURL)
+        if fileSystem.itemExists(at: directoryURL) {
+            try fileSystem.removeItem(at: directoryURL)
+        }
     }
 }
